requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
numpy
requests
pdfplumber
python-dotenv
pydantic

##Use Below command to install dependencies
##pip install -r requirements.txt
